BufferedInputStream / move code
